Re-factor profile.py within the suite app

+++ suite/views/profile.py
@@ -10,10 +10,16 @@ from django.shortcuts import (
     reverse,
     HttpResponse
 )
+from django.core import exceptions
 from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 from suite.models import Folio, Profile
-from suite.functions import id_has_been_provided, sort_by_id
+from suite.functions import (
+    id_has_been_provided,
+    sort_by_id,
+    user_is_author_of_snippet,
+    user_is_author_of_folio
+)
 from suite.forms import FolioProfileForm, FolioProfileCurrentAndFutureGoalForm
 
 
@@ -25,28 +31,20 @@ def edit_folio_profile(request, folio_id=None):
     """
 
     if id_has_been_provided(folio_id):
-
         folio = get_object_or_404(Folio, pk=folio_id)
 
-        # Get the user's profiles
         profiles = list(Profile.objects.filter(
             author_id=request.user
         ))
 
-        # For each profile, attach a form to the object
-        # And assess whether profile is attached to form
         for profile in profiles:
             profile.form = FolioProfileForm(
                 instance=profile,
                 prefix=f"profile-{profile.id}"
             )
-
-            # Set is_attached to true if folio exists
-            # in the profiles list of folios
             profile.is_attached = profile.folios.filter(pk=folio_id).exists()
 
         create_profile_form = FolioProfileForm()
-
         current_and_future_goal_form = FolioProfileCurrentAndFutureGoalForm(
             instance=folio
         )
@@ -57,11 +55,9 @@ def edit_folio_profile(request, folio_id=None):
             "current_and_future_goal_form": current_and_future_goal_form,
             "profiles": profiles
         }
-
         return render(request, "suite/edit_profile.html", context=context)
 
     else:
-        # If one hasn't been provided
         return redirect("select_folio")
 
 
@@ -71,36 +67,37 @@ def create_folio_profile(request, folio_id):
     Creates a new folio profile
     """
 
-    # Ensure the request made is a POST request
     if request.method == "POST":
-
-        # Create instance of profile form using form data
         form = FolioProfileForm(request.POST)
 
-        # If form is valid
         if form.is_valid():
-
-            # Partially save the form, as author_id
-            # will need to be provided
             profile = form.save(commit=False)
-
-            # Save current user as author of profile
             profile.author_id = request.user
-
-            # Fully save profile
             profile.save()
-
             messages.success(
                 request,
                 f"The {profile.profile_title} About Me has "
                 f"been successfully created."
             )
 
-            # Return to folio profile page using folio id
-            return redirect(
-                reverse("edit_folio_profile",
-                        kwargs={"folio_id": folio_id})
+        else:
+            messages.error(
+                request,
+                "Data posted was not valid "
+                "to create a new profile."
             )
+
+    else:
+        messages.error(
+            request,
+            "Data should be posted when "
+            "attempting to create a new profile."
+        )
+
+    return redirect(
+        reverse("edit_folio_profile",
+                kwargs={"folio_id": folio_id})
+    )
 
 
 @login_required
@@ -109,28 +106,50 @@ def update_folio_profile(request, profile_id, folio_id):
     Updates an existing folio profile
     """
 
-    if request.method == "POST":
-        profile = get_object_or_404(Profile, pk=profile_id)
+    if user_is_author_of_snippet(request.user, 'profile', profile_id):
 
-        post = request.POST.copy()
-        for key, value in request.POST.items():
-            prefix_removed_name = key.replace(f"profile-{profile.id}-", "")
-            post[prefix_removed_name] = value
+        if request.method == "POST":
+            profile = get_object_or_404(Profile, pk=profile_id)
+            post = request.POST.copy()
+            for key, value in request.POST.items():
+                prefix_removed_name = key.replace(f"profile-{profile.id}-", "")
+                post[prefix_removed_name] = value
 
-        form = FolioProfileForm(post, instance=profile)
+            form = FolioProfileForm(post, instance=profile)
 
-        if form.is_valid():
-            form.save()
-            messages.success(
+            if form.is_valid():
+                form.save()
+                messages.success(
+                    request,
+                    f"The {profile.profile_title} About Me has "
+                    f"been successfully updated."
+                )
+
+            else:
+                messages.error(
+                    request,
+                    "Data posted was not valid "
+                    "to update a profile."
+                )
+        else:
+            messages.error(
                 request,
-                f"The {profile.profile_title} About Me has "
-                f"been successfully updated."
+                "Data should be posted when "
+                "attempting to update a profile."
             )
 
         return redirect(
             reverse("edit_folio_profile",
                     kwargs={"folio_id": folio_id})
         )
+
+    else:
+        messages.error(
+            request,
+            "You cannot interact with "
+            "profile snippets that are not your own."
+        )
+        return redirect("view_library")
 
 
 @login_required
@@ -140,49 +159,55 @@ def update_profiles_attached_to_folio(request, folio_id):
     to the currently viewed folio
     """
 
-    if request.method == "POST":
-        # Retrieve list of project status's sorted by id
-        data = json.loads(request.body)
-        list_of_profiles = data['profiles']
-        list_of_profiles.sort(reverse=False, key=sort_by_id)
+    if user_is_author_of_folio(request.user, folio_id):
 
-        # Grab users profiles from DB using ID's provided
-        profiles_in_db = Profile.objects.filter(
-             id__in=[profile['id'] for profile in list_of_profiles]
-        )
+        if request.method == "POST":
+            data = json.loads(request.body)
+            list_of_profiles = data['profiles']
+            list_of_profiles.sort(reverse=False, key=sort_by_id)
 
-        # Get currently viewed folio
-        folio = get_object_or_404(Folio, pk=folio_id)
+            profiles_in_db = Profile.objects.filter(
+                id__in=[profile['id'] for profile in list_of_profiles]
+            )
+            folio = get_object_or_404(Folio, pk=folio_id)
 
-        # Iterate through profiles in db & status's given from user actions
-        for profile_in_db, profile_status in zip(profiles_in_db,
-                                                 list_of_profiles):
-            # Ensure ID's match
-            if profile_in_db.id == int(profile_status['id']):
-                # Add/remove folio or continue based on is_attached
-                # status & if the folio already exists in project folios
-                if profile_status['is_attached']:
-                    if profile_in_db.folios.filter(pk=folio_id).exists():
-                        continue
+            # Iterate through profiles in db & status's given from user actions
+            for profile_in_db, profile_status in zip(profiles_in_db,
+                                                     list_of_profiles):
+                # Ensure ID's match
+                if profile_in_db.id == int(profile_status['id']):
+                    # Add/remove folio or continue based on is_attached
+                    # status & if the folio already exists in project folios
+                    if profile_status['is_attached']:
+                        if profile_in_db.folios.filter(pk=folio_id).exists():
+                            continue
+                        else:
+                            profile_in_db.folios.add(folio)
                     else:
-                        profile_in_db.folios.add(folio)
+                        if profile_in_db.folios.filter(pk=folio_id).exists():
+                            profile_in_db.folios.remove(folio)
+                        else:
+                            continue
+
+                    messages.success(
+                        request,
+                        f"The profile snippets attached to {folio.name} "
+                        f"have been updated successfully."
+                    )
+
                 else:
-                    if profile_in_db.folios.filter(pk=folio_id).exists():
-                        profile_in_db.folios.remove(folio)
-                    else:
-                        continue
+                    print(
+                        "DEBUG: Profile id's didn't match, "
+                        "likely sorting error"
+                    )
 
-                messages.success(
-                    request,
-                    f"About Me profiles attached to {folio.name} "
-                    f"have been updated successfully."
-                )
+            return HttpResponse("OK")
 
-            else:
-                print("they didn't match, sorting error")
+        else:
+            raise exceptions.BadRequest
 
-        # Return an OK response
-        return HttpResponse("OK")
+    else:
+        raise exceptions.PermissionDenied
 
 
 @login_required
@@ -193,32 +218,49 @@ def update_current_and_future_goal(request, folio_id):
     profile tab in the suite.
     """
 
-    # Ensure the request made is a POST request
-    if request.method == "POST":
+    if user_is_author_of_folio(request.user, folio_id):
 
-        # Get current folio
-        folio_in_db = get_object_or_404(Folio, pk=folio_id)
+        if request.method == "POST":
+            folio = get_object_or_404(Folio, pk=folio_id)
 
-        # Save instance of profile form
-        form = FolioProfileCurrentAndFutureGoalForm(
-            request.POST,
-            instance=folio_in_db
-        )
-
-        # Save form if valid
-        if form.is_valid():
-            form.save()
-            messages.success(
-                request,
-                f"Current & future goals within {folio_in_db.name} "
-                f"have been updated successfully."
+            form = FolioProfileCurrentAndFutureGoalForm(
+                request.POST,
+                instance=folio
             )
 
-        # Return to folio profile page using folio id
+            if form.is_valid():
+                form.save()
+                messages.success(
+                    request,
+                    f"Current & future goals within {folio.name} "
+                    f"have been updated successfully."
+                )
+            else:
+                messages.error(
+                    request,
+                    "Data posted was not valid "
+                    "to update current project & future goal."
+                )
+        else:
+            messages.error(
+                request,
+                "Data should be posted when "
+                "attempting to update your "
+                "current project and future goal."
+            )
+
         return redirect(reverse(
             "edit_folio_profile",
             kwargs={"folio_id": folio_id}
         ))
+
+    else:
+        messages.error(
+            request,
+            "You cannot interact with content within "
+            "folios that are not your own."
+        )
+        return redirect("view_library")
 
 
 @login_required
@@ -227,23 +269,33 @@ def delete_folio_profile(request, profile_id, folio_id):
     Deletes a user's folio profile
     """
 
-    # Ensure the request made is a POST request
-    if request.method == "POST":
+    if user_is_author_of_snippet(request.user, 'profile', profile_id):
+    
+        if request.method == "POST":
+            profile_in_db = get_object_or_404(Profile, pk=profile_id)
+            profile_in_db.delete()
+            messages.success(
+                request,
+                f"The {profile_in_db.profile_title} About Me profile "
+                f"has been deleted successfully."
+            )
+        
+        else:
+            messages.error(
+                request,
+                "A delete command should "
+                "be sent as a POST request."
+            )
 
-        # Get the profile
-        profile_in_db = get_object_or_404(Profile, pk=profile_id)
-
-        # Delete the profile from the database
-        profile_in_db.delete()
-
-        messages.success(
-            request,
-            f"The {profile_in_db.profile_title} About Me profile "
-            f"has been deleted successfully."
-        )
-
-        # Return to folio profile page using folio id
         return redirect(
             reverse("edit_folio_profile",
                     kwargs={"folio_id": folio_id})
         )
+    
+    else:
+        messages.error(
+            request,
+            "You cannot delete profile snippets "
+            "that are not your own."
+        )
+        return redirect("view_library")
